fixed wakeup issues and level up issues

+++ Assignment 1/Main/LevelManager.cpp
@@ -56,7 +56,7 @@ int checkDifficulty() {
   }
 }
 
-long getFactor() {
+float getFactor() {
   return factor;
 }
